+++ R/TcGSAseq-package.R
@@ -6,11 +6,11 @@
 #'\tabular{ll}{
 #'Package: \tab tcgsaseq\cr
 #'Type: \tab Package\cr
-#'Version: \tab 1.8.1\cr
-#'Date: \tab 2018-12-06\cr
+#'Version: \tab 2.0.5\cr
+#'Date: \tab 2020-09-10\cr
 #'License:\tab \href{http://www.gnu.org/licenses/gpl-2.0.txt}{GPL-2}\cr
 #'}
-#'The two main functions of the \code{tcgsaseq} package are \code{\link{varseq}} and \code{\link{tcgsa_seq}}.
+#'The two main functions of the \code{tcgsaseq} package are \code{\link{varcompseq}} and \code{\link{tcgsa_seq}}.
 #'
 #'
 #'@author Boris P. Hejblum, Denis Agniel