+++ R/sp_weights.R
@@ -45,7 +45,7 @@
 #'
 #'@param transform a logical flag indicating whether values should be transformed to uniform
 #'for the purpose of local linear smoothing. This may be helpful if tail observations are sparse and
-#'the specified bandwidth gives suboptimal performance there. Default is \code{FALSE}.
+#'the specified bandwidth gives suboptimal performance there. Default is \code{TRUE}.
 #'
 #'@param verbose a logical flag indicating whether informative messages are printed
 #'during the computation. Default is \code{TRUE}.
@@ -80,7 +80,7 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
                        gene_based = FALSE,
                        bw = c("nrd", "ucv", "SJ", "nrd0", "bcv"),
                        kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight", "tricube", "cosine", "optcosine"),
-                       exact = FALSE, transform = FALSE,
+                       exact = FALSE, transform = TRUE,
                        verbose = TRUE,
                        na.rm = FALSE
 ){
@@ -131,6 +131,7 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
   mu <- xphi%*%B_ols
 
   sq_err <- (y_lcpm - mu)^2
+  lse <- log(sq_err)
   v <- colMeans(sq_err, na.rm = na.rm)
   mu_avg <- colMeans(mu, na.rm = na.rm)
 
@@ -165,11 +166,11 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
                    stop("unknown bandwidth rule: 'bw' argument must be among 'nrd0', 'nrd', 'ucv', 'bcv', 'SJ'"))
     }else{
       bw <- switch(bw,
-                   nrd0 = stats::bw.nrd0(as.vector(mu)),
-                   nrd = stats::bw.nrd(as.vector(mu)),
-                   ucv = stats::bw.ucv(as.vector(mu)),
-                   bcv = stats::bw.bcv(as.vector(mu)),
-                   SJ = stats::bw.SJ(as.vector(mu), method = "ste"),
+                   nrd0 = stats::bw.nrd0(as.vector(mu_x)),
+                   nrd = stats::bw.nrd(as.vector(mu_x)),
+                   ucv = stats::bw.ucv(as.vector(mu_x)),
+                   bcv = stats::bw.bcv(as.vector(mu_x)),
+                   SJ = stats::bw.SJ(as.vector(mu_x), method = "ste"),
                    stop("unknown bandwidth rule: 'bw' argument must be among 'nrd0', 'nrd', 'ucv', 'bcv', 'SJ'"))
     }
     if(verbose){
@@ -185,6 +186,7 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
   }
 
 
+  # choose kernel
   if(kernel == "gaussian"){
     kern_func <- function(x, bw){
       stats::dnorm(x, sd = bw)
@@ -234,6 +236,7 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
   }
 
 
+  # compute the weights
   if(gene_based){
 
     w <- function(x){
@@ -251,7 +254,7 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
       message("'exact' is TRUE: the computation may take up to a couple minutes...", "\n",
           "Set 'exact = FALSE' for quicker computation of the weights\n")
 
-      weights <- t(matrix(1/unlist(lapply(as.vector(mu), w)), ncol = n, nrow = p, byrow = FALSE))
+      weights <- t(matrix(1/unlist(lapply(as.vector(mu_x), w)), ncol = n, nrow = p, byrow = FALSE))
       if(sum(!is.finite(weights))>0){
         warning("At least 1 non finite weight. Try to increase the bandwith")
       }
@@ -270,15 +273,21 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
     }else if(sum(is.na(mu_x))>1){
       mu_x <- mu_x[-which(is.na(mu_x))]
       sq_err <- sq_err[-which(is.na(sq_err))]
+      lse <- lse[-which(is.na(lse))]
     }
-    smth <- KernSmooth::locpoly(x = c(mu_x), y = c(sq_err),
-                                degree = 1, kernel = kernel, bandwidth = bw)
-    w <- (1/stats::approx(reverse_trans(smth$x), smth$y, xout = mu, rule = 2)$y)
-    weights <- matrix(w, nrow(mu), ncol(mu))
+    smth <- KernSmooth::locpoly(x = c(mu_x), y = c(lse),
+                                degree = 2, kernel = kernel, bandwidth = bw)
+    w <- (1/exp(stats::approx(x = reverse_trans(smth$x), y = smth$y,
+                              xout = reverse_trans(mu_x), rule = 2)$y))
+    weights <- matrix(w, nrow(mu_x), ncol(mu_x))
+    if(sum(weights<0)>1){
+      stop("negative variance weights estimated: please contact the authors of the package")
+    }
   }
 
 
   if(doPlot){
+    inds <- list()
     if (gene_based) {
       o <- order(mu_avg, na.last = NA)
       plot_df <- data.frame("m_o" = mu_avg[o], "v_o" = v[o])
@@ -289,24 +298,40 @@ sp_weights <- function(y, x, phi, use_phi=TRUE, preprocessed = FALSE, doPlot = F
       }
       plot_df_lo <- data.frame("lo.x" = mu_avg[o], "lo.y" = kern_fit)
     } else {
-      inds <- sample(1:length(mu), size = 1000)
-      mu_s <- mu[inds]
-      ep_s <- sq_err[inds]
+      grid <- seq(from=min(mu_x), to= max(mu_x), length.out = 20)
+      n_mu_x <- length(mu_x)
+      for (i in 2:length(grid)){
+        possibles <- which(mu_x>=grid[i-1] & mu_x<=grid[i])
+        n.points <- max(2000, min(length(possibles)/20, 5000))
+        if(length(possibles)<n.points){
+          n.points <- length(possibles)
+        }
+        inds[[i]] <- sample(possibles, size = n.points)
+      }
+      inds <- unique(unlist(inds))
+      #inds <- mu_x#sample(1:length(mu_x), size = 1000)
+      mu_s <- reverse_trans(mu_x)[inds]
+      ep_s <- lse[inds]
       plot_df <- data.frame("m_o" = mu_s, "v_o" = ep_s)
       plot_df_lo <- data.frame("lo.x" = reverse_trans(smth$x), "lo.y" = smth$y)
     }
 
-
     #plot_df_lo_temp <- data.frame("lo.x" = mu_avg[o], "lo.y" = f_interp(mu_avg[o]))
-    ggp <- (ggplot(data = plot_df)
-            + geom_point(aes_string(x = "m_o", y = "v_o"), alpha = 0.45, color = "grey25", size = 0.5)
-            + theme_bw()
-            + xlab("Conditional mean")
-            + ylab("Variance")
-            + ggtitle("Mean-variance local regression non-parametric fit")
-            + geom_line(data = plot_df_lo, aes_string(x = "lo.x", y = "lo.y"), color = "blue", lwd = 1.4, lty = "solid", alpha = 0.8)
+    ggp <- (ggplot(data = plot_df) +
+            geom_point(aes_string(x = "m_o", y = "v_o"), alpha = 0.4, color = "grey25", size = 0.6) +
+            theme_bw() +
+            xlab("Observed (transformed) counts") +
+            ylab("log squared-error") +
+            ggtitle("Mean-variance local regression non-parametric fit") +
+            geom_line(data = plot_df_lo, aes_string(x = "lo.x", y = "lo.y"), color = "blue", lwd = 1.4, lty = "solid", alpha = 0.5)
             #+ geom_line(data = plot_df_lo_temp, aes(x = lo.x, y = lo.y), color = "red", lwd = 1, lty = 2)
     )
+    if(length(inds)>1){
+      ggp <- ggp +
+        ylim(range(lse)) +
+        xlim(range(reverse_trans(mu_x))) +
+        labs(subtitle = paste(length(inds), "subsampled points"))
+    }
     print(ggp)
   }