+++ R/varcompseq.R
@@ -36,7 +36,19 @@
 #'the variance component score test, either \code{"permutation"} or \code{"asymptotic"}.
 #'Default is \code{"permutation"}.
 #'
-#'@param n_perm the number of perturbations. Default is \code{1000}
+#'@param n_perm the number of perturbations. Default is \code{1000}.
+
+#'@param progressbar logical indicating whether a progress bar should be displayed
+#'when computing permutations (only in interactive mode).
+#'
+#'@param parallel_comp a logical flag indicating whether parallel computation
+#'should be enabled. Only Linux and MacOS are supported, this is ignored on Windows.
+#'Default is \code{TRUE}.
+#'
+#'@param nb_cores an integer indicating the number of cores to be used when
+#'\code{parallel_comp} is \code{TRUE}.
+#'Only Linux and MacOS are supported, this is ignored on Windows.
+#'Default is \code{parallel::detectCores() - 1}.
 #'
 #'@param preprocessed a logical flag indicating whether the expression data have
 #'already been preprocessed (e.g. log2 transformed). Default is \code{FALSE}, in
@@ -68,7 +80,7 @@
 #'@param transform a logical flag used for \code{"loclin"} weights, indicating whether values should be
 #'transformed to uniform for the purpose of local linear smoothing. This may be helpful if tail
 #'observations are sparse and the specified bandwidth gives suboptimal performance there.
-#'Default is \code{FALSE}.
+#'Default is \code{TRUE}.
 #'
 #'@param padjust_methods multiple testing correction method used if \code{genesets}
 #'is a list. Default is "BH", i.e. Benjamini-Hochberg procedure for controlling the FDR.
@@ -83,7 +95,7 @@
 #'@param homogen_traj a logical flag indicating whether trajectories should be considered homogeneous.
 #'Default is \code{FALSE} in which case trajectories are not only tested for trend, but also for heterogeneity.
 #'
-#'@param na.rm_varseq logical: should missing values in \code{y} (including
+#'@param na.rm_varcompseq logical: should missing values in \code{y} (including
 #'\code{NA} and \code{NaN}) be omitted from the calculations?
 #'Default is \code{FALSE}.
 #'
@@ -133,7 +145,7 @@
 #'x <- matrix(1, ncol=1, nrow=r)
 #'
 #'#run test
-#'res_genes <- varseq(exprmat=y, covariates=x, variables2test=t, sample_group=rep(1:ni, each=nr),
+#'res_genes <- varcompseq(exprmat=y, covariates=x, variables2test=t, sample_group=rep(1:ni, each=nr),
 #'                    which_test="asymptotic",
 #'                    which_weights="none", preprocessed=TRUE)
 #'mean(res_genes$pvals[, "rawPval"]>0.05)
@@ -149,24 +161,25 @@
 #'y.tilde <- b0 + b1*t + rnorm(r, sd = sigma)
 #'y <- t(matrix(rnorm(n*r, sd = sqrt(sigma*abs(y.tilde))), ncol=n, nrow=r) +
 #'       matrix(rep(y.tilde, n), ncol=n, nrow=r))
-#'res_genes <- varseq(exprmat=y, covariates=x, variables2test=t, sample_group=rep(1:ni, each=nr),
-#'                    which_weights="none", preprocessed=TRUE, n_perm=1000)
+#'res_genes <- varcompseq(exprmat=y, covariates=x, variables2test=t, sample_group=rep(1:ni, each=nr),
+#'                    which_weights="none", preprocessed=TRUE)
 #'summary(res_genes$pvals)
 #'}
 #'@export
-varseq <- function(exprmat, covariates, variables2test,
+varcompseq <- function(exprmat, covariates, variables2test,
                    sample_group = NULL, weights_var2test_condi = TRUE,
                    cov_variables2test_eff = diag(ncol(variables2test)),
                    which_test = c("permutation", "asymptotic"),
                    which_weights = c("loclin", "voom", "none"),
-                   n_perm = 1000,
+                   n_perm = 1000, progressbar = TRUE, parallel_comp = TRUE,
+                   nb_cores = parallel::detectCores() - 1,
                    preprocessed = FALSE, doPlot = TRUE, gene_based_weights = FALSE,
                    bw = "nrd",
                    kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight", "tricube", "cosine", "optcosine"),
-                   exact = FALSE, transform = FALSE,
+                   exact = FALSE, transform = TRUE,
                    padjust_methods = c("BH", "BY", "holm", "hochberg", "hommel", "bonferroni"),
                    lowess_span = 0.5,
-                   na.rm_varseq = TRUE,
+                   na.rm_varcompseq = TRUE,
                    homogen_traj = FALSE){
 
   return(tcgsa_seq(y = exprmat, x = covariates, phi = variables2test,
@@ -176,7 +189,8 @@ varseq <- function(exprmat, covariates, variables2test,
                    Sigma_xi = cov_variables2test_eff,
                    which_test = which_test,
                    which_weights = which_weights,
-                   n_perm = n_perm,
+                   n_perm = n_perm, progressbar = progressbar,
+                   parallel_comp = parallel_comp, nb_cores = nb_cores,
                    preprocessed = preprocessed, doPlot = doPlot, gene_based_weights = gene_based_weights,
                    bw = bw,
                    kernel = kernel,
@@ -184,7 +198,7 @@ varseq <- function(exprmat, covariates, variables2test,
                    padjust_methods = padjust_methods,
                    lowess_span = lowess_span,
                    homogen_traj = homogen_traj,
-                   na.rm_tcgsaseq = na.rm_varseq,
+                   na.rm_tcgsaseq = na.rm_varcompseq,
                    verbose = FALSE))
 
 }