+++ NEWS.md
@@ -1,5 +1,17 @@
 # News about the `tcgsaseq` R package
 
+### Main changes in Version 2.0.1 (2020-09-09)
+ * renaming `varseq()` function into `varcompseq` (following trademark infringement notification from Golden Helix, Inc)
+ * indicating soft-deprecation of `tcgsaseq`in favor of the `dearseq` Bioconductor package
+
+### Main changes in Version 2.0.0 (2019-08-02):
+ * `transform = TRUE` is now the default
+ * `loc-lin` weights are now computed on unconditional (transformed) counts. They should be more robust this way to outlier genes
+
+### Main changes in Version 1.9.0 (2019-07-18):
+ * (partial) parallel support for permutations
+ * progressbar for permutations in interactive mode
+ * bug fix for `voom`-like weights
 
 ### Main changes in Version 1.8.1 (2018-12-06):
 * perm_pe function for exact p-values and accurate FDR control when using permutation test.